Created MVP of app
Made most of the app's UI. Just need text fields for the user's name and residence hall (which will autofill as the user types), and I should be good to go!

This project idea was inspired by a friend who potentially needs an app that can take food orders. This will be a fun personal project!