Fixed issues with React Navigation. Polished menu a bit
Accomplished two big things. One is that I successfully installed React Navigation (took a lot more effort than it should've), and I added some nice colors to the app. The colors are likely to change, but that's not a big deal for now.

This commit is unusually large because I deleted the node_modules folder and the package-lock (reinstalling them after of course) to fix some issues I had with installation. A very productive day!